# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/user-settings": {
      "patch": {
        "operationId": "updateUserSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/acknowledge-welcome": {
      "patch": {
        "operationId": "acknowledgeWelcome",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeWelcomeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
...
    "/api/apps/{urn}": {
      "get": {
        "operationId": "getApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppDto"
                }
              }
            }
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
    "/api/apps/{urn}/update-diff": {
      "get": {
        "operationId": "getAppUpdateDiff",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUpdateDiffDto"
                }
              }
            }
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
...
    "/api/app-lifecycle/{urn}/install": {
      "post": {
        "operationId": "installApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppFormBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/start": {
      "post": {
        "operationId": "startApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/stop": {
      "post": {
        "operationId": "stopApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/restart": {
      "post": {
        "operationId": "restartApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/uninstall": {
      "delete": {
        "operationId": "uninstallApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UninstallAppBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/reset": {
      "post": {
        "operationId": "resetApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/update": {
      "patch": {
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
    "/api/app-lifecycle/{urn}/update-config": {
      "patch": {
        "operationId": "updateAppConfig",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppFormBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "AppLifecycle"
        ]
      }
    },
...
  },
  "info": {
    "title": "Runtipi API",
    "description": "API specs for Runtipi",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserSettingsBody": {
        "type": "object",
        "properties": {
          "advancedSettings": {
            "type": "boolean"
          },
          "allowAutoThemes": {
            "type": "boolean"
          },
          "allowErrorMonitoring": {
            "type": "boolean"
          },
          "appDataPath": {
            "type": "string"
          },
          "appsRepoUrl": {
            "type": "string",
            "format": "uri"
          },
          "demoMode": {
            "type": "boolean"
          },
          "dnsIp": {
            "type": "string",
            "pattern": "^((?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])[.]){3}(?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$",
            "format": "ipv4"
          },
          "domain": {
            "type": "string"
          },
          "eventsTimeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string",
                "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
              }
            ]
          },
          "experimental_insecureCookie": {
            "type": "boolean"
          },
          "forwardAuthUrl": {
            "type": "string",
            "format": "uri"
          },
          "guestDashboard": {
            "type": "boolean"
          },
          "internalIp": {
            "type": "string",
            "pattern": "^((?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])[.]){3}(?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$",
            "format": "ipv4"
          },
          "listenIp": {
            "type": "string",
            "pattern": "^((?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])[.]){3}(?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$",
            "format": "ipv4"
          },
          "localDomain": {
            "type": "string"
          },
          "logLevel": {
            "type": "string",
            "enum": [
              "debug",
              "error",
              "info",
              "warn"
            ]
          },
          "persistTraefikConfig": {
            "type": "boolean"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string",
                "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
              }
            ]
          },
          "postgresPort": {
            "type": "integer",
            "maximum": 65535,
            "minimum": 0
          },
          "sslPort": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string",
                "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
              }
            ]
          },
          "themeBase": {
            "type": "string"
          },
          "themeColor": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          }
        }
      },
      "AcknowledgeWelcomeBody": {
        "type": "object",
        "properties": {
          "allowErrorMonitoring": {
            "type": "boolean"
          }
        },
        "required": [
          "allowErrorMonitoring"
        ]
      },
...
      "GetAppDto": {
        "type": "object",
        "properties": {
          "info": {
            "type": "object",
            "properties": {
              "author": {
                "type": "string"
              },
              "available": {
                "type": "boolean"
              },
              "categories": {
                "type": "array",
                "items": {
                  "enum": [
                    "ai",
                    "automation",
                    "books",
                    "data",
                    "development",
                    "featured",
                    "finance",
                    "gaming",
                    "media",
                    "music",
                    "network",
                    "photography",
                    "security",
                    "social",
                    "utilities"
                  ]
                }
              },
              "created_at": {
                "type": "integer",
                "minimum": 0
              },
              "deprecated": {
                "type": "boolean"
              },
              "description": {
                "type": "string"
              },
              "dynamic_config": {
                "type": "boolean"
              },
              "exposable": {
                "type": "boolean"
              },
              "force_expose": {
                "type": "boolean"
              },
              "force_pull": {
                "type": "boolean"
              },
              "form_fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "env_variable": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "required": {
                      "type": "boolean"
                    },
                    "type": {
                      "enum": [
                        "boolean",
                        "email",
                        "fqdn",
                        "fqdnip",
                        "ip",
                        "number",
                        "password",
                        "random",
                        "text",
                        "url"
                      ]
                    },
                    "default": {
                      "anyOf": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "string"
                        },
                        {
                          "type": "boolean"
                        }
                      ]
                    },
                    "encoding": {
                      "enum": [
                        "base64",
                        "hex"
                      ]
                    },
                    "hint": {
                      "type": "string"
                    },
                    "max": {
                      "type": "number"
                    },
                    "min": {
                      "type": "number"
                    },
                    "options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string"
                          },
                          "value": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "label",
                          "value"
                        ]
                      }
                    },
                    "pattern_error": {
                      "type": "string"
                    },
                    "placeholder": {
                      "type": "string"
                    },
                    "regex": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "env_variable",
                    "label",
                    "required",
                    "type"
                  ]
                }
              },
              "generate_vapid_keys": {
                "type": "boolean"
              },
              "https": {
                "type": "boolean"
              },
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "no_gui": {
                "type": "boolean"
              },
              "short_desc": {
                "type": "string"
              },
              "source": {
                "type": "string"
              },
              "tipi_version": {
                "type": "number"
              },
              "updated_at": {
                "type": "integer",
                "minimum": 0
              },
              "urn": {
                "type": "string"
              },
              "version": {
                "type": "string"
              },
              "gid": {
                "type": "number"
              },
              "min_tipi_version": {
                "type": "string"
              },
              "port": {
                "type": "number",
                "maximum": 65535,
                "minimum": 1
              },
              "supported_architectures": {
                "type": "array",
                "items": {
                  "enum": [
                    "amd64",
                    "arm64"
                  ]
                }
              },
              "uid": {
                "type": "number"
              },
              "url_suffix": {
                "type": "string"
              },
              "website": {
                "type": "string"
              }
            },
            "required": [
              "author",
              "available",
              "categories",
              "created_at",
              "deprecated",
              "description",
              "dynamic_config",
              "exposable",
              "force_expose",
              "force_pull",
              "form_fields",
              "generate_vapid_keys",
              "https",
              "id",
              "name",
              "no_gui",
              "short_desc",
              "source",
              "tipi_version",
              "updated_at",
              "urn",
              "version"
            ]
          },
          "metadata": {
            "type": "object",
            "properties": {
              "latestVersion": {
                "type": "number"
              },
              "localSubdomain": {
                "type": "string"
              },
              "hasCustomConfig": {
                "type": "boolean"
              },
              "latestDockerVersion": {
                "type": "string"
              },
              "minTipiVersion": {
                "anyOf": [
                  {
                    "type": "string"
                  }
                ]
              }
            },
            "required": [
              "latestVersion",
              "localSubdomain"
            ]
          },
          "app": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "domain": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "exposed": {
                    "type": "boolean"
                  },
                  "exposedLocal": {
                    "type": "boolean"
                  },
                  "id": {
                    "type": "number"
                  },
                  "isVisibleOnGuestDashboard": {
                    "type": "boolean"
                  },
                  "openPort": {
                    "type": "boolean"
                  },
                  "pendingRestart": {
                    "type": "boolean"
                  },
                  "port": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "enum": [
                      "backing_up",
                      "installing",
                      "missing",
                      "resetting",
                      "restarting",
                      "restoring",
                      "running",
                      "starting",
                      "stopped",
                      "stopping",
                      "uninstalling",
                      "updating"
                    ]
                  },
                  "version": {
                    "type": "number"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "enableAuth": {
                    "type": "boolean"
                  },
                  "localSubdomain": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "domain",
                  "exposed",
                  "exposedLocal",
                  "id",
                  "isVisibleOnGuestDashboard",
                  "openPort",
                  "pendingRestart",
                  "port",
                  "status",
                  "version"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "info",
          "metadata"
        ]
      },
      "GetUpdateDiffDto": {
        "type": "object",
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "new": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "current",
          "new"
        ]
      },
...
      "AppFormBody": {
        "type": "object",
        "properties": {
          "skipEnv": {
            "type": "boolean",
            "default": false
          },
          "skipPull": {
            "type": "boolean",
            "default": false
          },
          "skipRun": {
            "type": "boolean",
            "default": false
          },
          "openPort": {
            "type": "boolean",
            "default": true
          },
          "domain": {
            "type": "string"
          },
          "enableAuth": {
            "type": "boolean"
          },
          "exposed": {
            "type": "boolean"
          },
          "exposedLocal": {
            "type": "boolean"
          },
          "isVisibleOnGuestDashboard": {
            "type": "boolean"
          },
          "localSubdomain": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9-]{1,63}$"
          },
          "port": {
            "type": "number",
            "maximum": 65535,
            "minimum": 1024
          }
        }
      },
      "LifecycleRequestDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string",
                "description": "a UUID",
                "format": "uuid",
                "pattern": "[\\da-f]{8}-[\\da-f]{4}-[1-8][\\da-f]{3}-[89ab][\\da-f]{3}-[\\da-f]{12}"
              },
              {
                "const": "00000000-0000-0000-0000-000000000000",
                "description": "a UUID",
                "format": "uuid"
              },
              {
                "const": "ffffffff-ffff-ffff-ffff-ffffffffffff",
                "description": "a UUID",
                "format": "uuid"
              }
            ],
            "description": "a UUID",
            "format": "uuid"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "UninstallAppBody": {
        "type": "object",
        "properties": {
          "removeBackups": {
            "type": "boolean"
          }
        },
        "required": [
          "removeBackups"
        ]
      },
      "UpdateAppBody": {
        "type": "object",
        "properties": {
          "performBackup": {
            "type": "boolean"
          }
        },
        "required": [
          "performBackup"
        ]
      },
...
    }
  }
}